Move Billing setup check to billing class
